Should Library be a location? Fixes #326
