+++ wod/models/characters/mage/mage.py
@@ -8,7 +8,7 @@ from django.urls import reverse
 from core.models import Model
 from core.utils import add_dot, weighted_choice
 from wod.models.characters.human import Human
-from wod.models.items.mage import Library
+from wod.models.locations.mage.library import Library
 from wod.models.locations.mage.nodes import Node
 
 from .faction import MageFaction
