+++ wod/models/locations/mage/__init__.py
@@ -1,3 +1,5 @@
 from .chantry import Chantry
+from .library import Library
 from .nodes import Node, NodeMeritFlaw, NodeMeritFlawRating, NodeResonanceRating
+from .realms import HorizonRealm
 from .sector import Sector
