+++ wod/tests/locations/mage.py
@@ -14,8 +14,8 @@ from wod.models.characters.mage import (
     Resonance,
 )
 from wod.models.characters.mage.utils import ABILITY_LIST, SPHERE_LIST
-from wod.models.items.mage import Library
-from wod.models.locations.mage import Chantry, Node, NodeMeritFlaw
+from wod.models.items.mage import Grimoire
+from wod.models.locations.mage import Chantry, Library, Node, NodeMeritFlaw
 from wod.tests.characters.mage import mage_setup
 from wod.tests.items.mage import grimoire_setup
 
@@ -281,14 +281,14 @@ class TestChantry(TestCase):
         self.assertFalse(self.chantry.has_library())
         self.chantry.create_library()
         self.assertTrue(self.chantry.has_library())
-        self.assertEqual(self.chantry.library.num_books(), 0)
+        self.assertEqual(self.chantry.chantry_library.num_books(), 0)
         self.chantry.library_rating = 4
         self.chantry.create_library()
         self.assertTrue(self.chantry.has_library())
-        self.assertEqual(self.chantry.library.num_books(), 4)
+        self.assertEqual(self.chantry.chantry_library.num_books(), 4)
 
     def test_set_library(self):
-        library = Library.objects.create(name="Test Library")
+        library = Library.objects.create(name="Test Library", rank=0)
         self.assertFalse(self.chantry.has_library())
         self.chantry.set_library(library)
         self.assertTrue(self.chantry.has_library())
@@ -420,3 +420,44 @@ class TestChantryDetailView(TestCase):
     def test_chantry_detail_view_templates(self):
         response = self.client.get(f"/wod/locations/{self.location.id}/")
         self.assertTemplateUsed(response, "wod/locations/mage/chantry/detail.html")
+
+
+class TestLibrary(TestCase):
+    def setUp(self):
+        self.grimoire = Grimoire()
+        grimoire_setup()
+        self.library = Library.objects.create(name="Test Library")
+
+    def test_add_book(self):
+        g = Grimoire.objects.create(name="Book To Add")
+        g.random()
+        count = self.library.num_books()
+        self.assertTrue(self.library.add_book(g))
+        self.assertEqual(self.library.num_books(), count + 1)
+
+    def test_has_books(self):
+        self.library.rank = 3
+        self.assertEqual(self.library.books.count(), 0)
+        self.library.books.add(Grimoire.objects.create(name="Test Grimoire 1", rank=1))
+        self.library.books.add(Grimoire.objects.create(name="Test Grimoire 2", rank=2))
+        self.library.books.add(Grimoire.objects.create(name="Test Grimoire 3", rank=3))
+        self.assertEqual(self.library.books.count(), 3)
+
+    def test_increase_library_rating(self):
+        self.assertEqual(self.library.num_books(), 0)
+        self.library.increase_rank()
+        self.library.increase_rank()
+        self.assertEqual(self.library.num_books(), 2)
+
+
+class TestLibraryDetailView(TestCase):
+    def setUp(self):
+        self.library = Library.objects.create(name="Test Library")
+
+    def test_library_detail_view_status_code(self):
+        response = self.client.get(f"/wod/locations/{self.library.id}/")
+        self.assertEqual(response.status_code, 200)
+
+    def test_library_detail_view_template(self):
+        response = self.client.get(f"/wod/locations/{self.library.id}/")
+        self.assertTemplateUsed(response, "wod/items/mage/library/detail.html")
